requirements.txt
altair==6.0.0
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///C:/Users/task_177249130442320/croot/anaconda-cloud-auth-split_1772491372848/work
anaconda-cli-base @ file:///C:/Users/task_177038553276457/croot/anaconda-cli-base_1770385908202/work
annotated-types @ file:///C:/miniconda3/conda-bld/annotated-types_1761745107361/work
anyio @ file:///C:/miniconda3/conda-bld/anyio_1758622433823/work
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.4.0
blinker==1.9.0
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961374486/work
cachetools==7.0.5
certifi @ file:///C:/Users/task_176765924055489/croot/certifi_1767659363027/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792955/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744975868/work
click @ file:///C:/miniconda3/conda-bld/click_1764332372183/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
conda @ file:///C:/Users/task_177140157604822/croot/conda_1771402255672/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///C:/b/abs_c9yejtsx9g/croot/conda-anaconda-tos_1755123332641/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///C:/miniconda3/conda-bld/conda-package-handling_1762366515145/work
conda_package_streaming @ file:///C:/miniconda3/conda-bld/conda-package-streaming_1762361689679/work
cryptography @ file:///C:/Users/task_177247529657270/croot/cryptography-split_1772475408670/work
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
frozendict @ file:///C:/miniconda3/conda-bld/frozendict_1761750728192/work
gitdb==4.0.12
GitPython==3.1.46
h11 @ file:///C:/miniconda3/conda-bld/h11_1761931384726/work
httpcore @ file:///C:/b/abs_25m7_xthp6/croot/httpcore_1748526065845/work
httpx @ file:///C:/miniconda3/conda-bld/httpx_1760447475275/work
idna @ file:///C:/miniconda3/conda-bld/idna_1761912043388/work
jaraco.classes @ file:///C:/b/abs_6erueoob1v/croot/jaraco.classes_1755516340851/work
jaraco.context @ file:///C:/Users/task_176947773609372/croot/jaraco.context_1769477902330/work
jaraco.functools @ file:///C:/Users/task_176838914677671/croot/jaraco.functools_1768389438594/work
Jinja2==3.1.6
jiter==0.13.0
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer @ file:///C:/b/abs_73u73l7pl9/croot/jsonpointer_1753788460913/work
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring @ file:///C:/miniconda3/conda-bld/keyring_1763637203252/work
libmambapy @ file:///C:/miniconda3/conda-bld/mamba-split_1763111608356/work/libmambapy
markdown-it-py @ file:///C:/Users/task_176700171154409/croot/markdown-it-py_1767001834659/work
MarkupSafe==3.0.3
mdurl @ file:///C:/miniconda3/conda-bld/mdurl_1758552277768/work
menuinst @ file:///C:/miniconda3/conda-bld/menuinst_1765382397455/work
more-itertools @ file:///C:/miniconda3/conda-bld/more-itertools_1761121564395/work
msgpack @ file:///C:/b/abs_4b3t4uhz3r/croot/msgpack-python_1750958631084/work
narwhals==2.18.0
numpy==2.4.3
openai==2.28.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049099114/work
pandas==2.3.3
pillow==12.1.1
pkce @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pkce_1729049216383/work
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356623609/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
protobuf==6.33.5
pyarrow==23.0.1
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
pydantic @ file:///C:/miniconda3/conda-bld/pydantic_1764083582233/work
pydantic-settings @ file:///C:/miniconda3/conda-bld/pydantic-settings_1764165236385/work
pydantic_core @ file:///C:/miniconda3/conda-bld/pydantic-core_1764009799098/work
pydeck==0.9.1
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431428918/work
PyJWT @ file:///C:/miniconda3/conda-bld/pyjwt_1764332257117/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030965/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///C:/Users/task_176855922952874/croot/python-dotenv_1768559755085/work
pytz==2026.1.post1
pywin32-ctypes @ file:///C:/Users/task_176829651200310/croot/pywin32-ctypes_1768296624234/work
readchar @ file:///C:/miniconda3/conda-bld/readchar_1760613474723/work
referencing==0.37.0
requests @ file:///C:/miniconda3/conda-bld/requests_1762359611326/work
rich @ file:///C:/miniconda3/conda-bld/rich_1760375661587/work
rpds-py==0.30.0
ruamel.yaml @ file:///C:/miniconda3/conda-bld/ruamel.yaml_1762536064547/work
ruamel.yaml.clib @ file:///C:/miniconda3/conda-bld/ruamel.yaml.clib_1762530094515/work
semver @ file:///C:/miniconda3/conda-bld/semver_1761903323755/work
setuptools==80.10.2
shellingham @ file:///C:/miniconda3/conda-bld/shellingham_1761912227081/work
six==1.17.0
smmap==5.0.3
sniffio @ file:///C:/miniconda3/conda-bld/sniffio_1764329222593/work
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tomli @ file:///C:/Users/task_176829653470725/croot/tomli_1768296655778/work
tomlkit @ file:///C:/miniconda3/conda-bld/tomlkit_1762896656167/work
tornado==6.5.5
tqdm @ file:///C:/Users/task_177139862118811/croot/tqdm_1771398768256/work
truststore @ file:///C:/miniconda3/conda-bld/truststore_1762521027919/work
typer==0.20.0
typer-slim==0.20.0
typing-inspection @ file:///C:/miniconda3/conda-bld/typing-inspection_1760614188477/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.3
urllib3 @ file:///C:/Users/task_176781035981662/croot/urllib3_1767810483132/work
watchdog==6.0.0
wheel==0.46.3
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278300/work
zstandard @ file:///C:/miniconda3/conda-bld/zstandard_1758189089298/work
